Add parameter to allow reordering due to memop clustering
Reordering based on the sort order of the MemOpInfo array was disabled
in <https://reviews.llvm.org/D72706>, but it's not clear this is
desirable for all targets. It also makes it more difficult to compare
the incremental benefit of enabling load clustering in the selectiondag
scheduler as well as the machinescheduler, as the sdag scheduler does
seem to allow this reordering.