[mlir] Start rewrite tool
Initial commit of a tool to help in textual rewrites of .mlir files. This tool
builds of of AsmParserState and is rather simple. Took some inspiration from
when I used clang's AST rewrites where I'd often treat it as a "localizing"
regex applicator in fallback cases, and started with that as functionality.
There though, one does have access to the lower level info than here, but still
a step up over sed over entire file.

This aims to be helpful (e.g., rewrite syntax including best effort inside
comments) rather than bulletproof tool. It may even be better suited under
utils than tools. And most of the rewrites would be rather short lived and
might never make it upstream (while the helpers of those rewrites may for
future rewrites).

The layering at the moment is not ideal as it is reusing the RewriteBuffer
class from clang's rewrite engine. So only optionally enabling where clang is
also enable. There doesn't seem to be anything clang specific there (the dep
does pull in more dependencies than ideal, but leaving both refactorings).

Additionally started it as a single file to prototype more easily, planning to
refactor later to include and libs for out of file usage.