[DirectX] Start documenting DXIL Resource handling
This adds a new document about DXIL Resource Handling. I've attempted
to describe here how we intend to use TargetExtTypes to represent
resources in LLVM IR and the various intrinsics we'll need to lower
these through LLVM to DXIL.

There are a number of TODOs in the document, but I think it makes
sense to publish as is and evolve it as we start implementing this.