[pstl] Avoid reusing moved-from iterators in PSTL tests
The reverse_invoker utility for PSTL tests uses forwarding references
for all parameters, but some of those parameters get forwarded to move
constructors which then leave the objects in a moved-from state. When
the parameters are forwarded a second time that results in making new
copies of moved-from iterators.  For libstdc++ debug mode iterators, the
moved-from state is singular, which means copying them will abort at
runtime.

The fix is to make copies of iterator arguments instead of forwarding
them. Doing this in reverse_invoker::operator() is sufficient to fix the
bug. It's OK to forward in invoke_on_all_iterator_types::operator() and
invoke_on_all_policies::operator() because those forwarded arguments do
not get copied until they're passed to reverse_invoker, which now copies
them as lvalues.

Fixes #80126