[VPlan] Delay adding canonical IV increment and exit branches.
This patch delays adding canonical IV increments, computing the next
active lane mask and the branch recipes to exit the vector loop.

During initial construction of a VPlan, only add the canonical IV phi
and active-lane-mask phis if needed. Similarly, do not add the branches
to exit the loop initially. Computing the next IV value, the next
active-lane-mask or the exit branches are details that are only needed
to simplify codegen (execute of the individual recipes).

This makes the inital VPlans more abstract (and simpler), in that we
initially leave out some details. Initial VPlans still have the canonical
induction recipe, which provides the value of the canonical induction at
every iteration, but we leave out the detail of how it is computed, which
is not needed until code generation.

Similar reasoning applies to active lane mask increments and branch
recipes. The vector loop region initially abstractly models the loop
processing all vector iterations, without specifying how exactly exiting
the region is handled. Again these details are only needed for code
generation.

To introduce the missing pieces and complete the abstract VPlan, a new
prepareToExecute transform is added and run just before exiting the
VPlan.

This is an attempt to further employ gradual lowering, as outlined in
https://llvm.org/devmtg/2023-10/slides/techtalks/Hahn-VPlan-StatusUpdateAndRoadmap.pdf
and already applied for replicate region handling.