[lldb] Add ability to show enum as name and value at the same time
When an enum is used to represent certain data it can be useful to know
its name and the value of it. For instance, register fields are
often set in source code as numbers, but in the debugger you'd like
to see the meaning as well.

(lldb) register read fpcr
    fpcr = 0x00000000
         = (... RMode = RN (0), ...)

Often you do just want the meaning but the value saves you having
to manually decode it if you want to confirm what your source code
has done, or try to replicate the current state in your source code.

I have not added tests for this because right now the use case for
this is registers and those will have their own test cases to cover
this. If we decide to expose this to formatters then this will need
more testing.