[CMake] Do not set CMP0116 explicitly to old
CMP0116 was originally set to old to get rid of warnings. However, this
behavior is now set to new by default with the minimum CMake version that LLVM
requires so does not produce any warnings, and setting it explicitly to old
does produce a warning in newer CMake versions. Due to these reasons, remove
this check for now.

Splitting off from removing the CMP0114 check just in case something breaks.

Partially fixes #83727.